+++ FluidTest.SampleTests/TestCases/TestData/UdpCoreFrameworkTests/Config/TestFullLoad.json
@@ -1,0 +1,28 @@
+{
+    "DF_Select_Fields_Transform": {
+        "Run": false,
+        "FieldsToInclude": null,
+        "FieldsToExclude": null,
+        "FieldsToNull": null,
+        "Mode": null
+    },
+    "ApprovalRequired": false,
+    "DF_SCD2_Delta_JSON":{
+        "Run" : true,
+        "TrackDeletions": true,
+        "SNKInsertPK": "PersonId",
+        "SNKUpdateSK": "id",
+        "HashColumns": "Name,PersonId,Age,Friends",
+        "SNKActiveFlagName": "Active",
+        "SNKDeletedFlagName": "Deleted",
+        "SNKHashName": "RowKeyHash",
+        "EntityName": "TestFullLoad",
+        "DeltaFolder": "PROCESSED/COMPLETED/TestFullLoad",
+        "DeletesLoadType": "FullLoad",
+        "SRCDeletedFlagName": null
+    },
+  "CosmosDB": {
+    "ContainerName": "TestFullLoad",
+    "CosmosDbName": "UdpTesting"
+  }
+}
