+++ FluidTest.SampleTests/TestCases/TestData/UdpCoreFrameworkTests/Data/TestFullLoad-delete.json
@@ -1,0 +1,28 @@
+[
+  {
+    "Name": "Markus",
+    "PersonId": 1,
+    "Age": 38,
+    "Friends": [
+      {
+        "Name": "Jim",
+        "Age": 32
+      },
+      {
+        "Name": "Leida",
+        "Age": 33
+      }
+    ]
+  },
+  {
+    "Name": "Daniel",
+    "PersonId": 2,
+    "Age": 38,
+    "Friends": [
+      {
+        "Name": "Mark",
+        "Age": 66
+      }
+    ]
+  }
+]
